requirements.txt
# Biznes Miyasi Bot — Dependencies
aiogram==3.13.1
aiosqlite==0.20.0
aiohttp==3.10.11
python-dotenv==1.0.1
fpdf2==2.8.2
openpyxl==3.1.5
